+++ api_docs/changelog.md
@@ -20,6 +20,11 @@ format used by the Zulip server that they are interacting with.
 
 ## Changes in Zulip 9.0
 
+**Feature level 257**
+* [`PATCH /messages/{message_id}`](/api/update-message),
+  [`PATCH /scheduled_messages/<int:scheduled_message_id>`](/api/update-scheduled-message):
+  These endpoints now return the detached_files within the response after editing the message.
+
 **Feature level 256**
 
 * [`GET /events`](/api/get-events): Stream update events with a new